requirements.txt
opencv-python
h5py
Pillow
requests
psutil
imageio
tensorflow
keras
numpy
imutils
tk